+++ common/src/main/java/net/spell_engine/config/ClientConfig.java
@@ -14,6 +14,8 @@ public class ClientConfig implements ConfigData {
 
     @ConfigEntry.Gui.Tooltip
     public boolean highlightTarget = true;
+    @ConfigEntry.Gui.Tooltip
+    public boolean stickyTarget = true;
 
 //    @ConfigEntry.Gui.Tooltip
 //    public boolean useMagicColorForHighlight = true;