+++ CHANGELOG.md
@@ -1,6 +1,7 @@
 # 0.9.5
 
 - Add spell power caching
+- Add sticky targets
 - Fix spell projectiles knocking back targets at incorrect angle
 
 # 0.9.4